combine separation tests
combined separation tests, because classify_audio would fail when other models ran on GPU
